+++ R/pkonfound.R
@@ -30,6 +30,8 @@
 #' @param c the number of cases in the treatment group showing unsuccessful results (2x2 table model).
 #' @param d the number of cases in the treatment group showing successful results (2x2 table model).
 #' @param two_by_two_table a table (matrix, data.frame, tibble, etc.) from which \code{a}, \code{b}, \code{c}, and \code{d} can be extracted.
+#' @param replace_stu score of the hypothetical average student who replaces the original student.
+#' @param peer_effect_pi proportion of students exerting peer effects on the others.  
 #' @param test specifies whether to use Fisher's Exact Test (\code{"fisher"}) or a chi-square test (\code{"chisq"}); defaults to \code{"fisher"}.
 #' @param to_return specifies the output format: \code{"print"} (default) to display output, \code{"plot"} for a plot, or \code{"raw_output"} to return a data.frame for further analysis.
 
@@ -52,6 +54,11 @@
 #' \strong{2x2 Table Model (Non-linear)}
 #' \itemize{
 #'   \item a, b, c, d, two_by_two_table, test, replace, switch_trm
+#' }
+#' 
+#' \strong{VAM model (beta)}
+#' \itemize{
+#'   \item est_eff, replace_stu, n_obs, eff_thr, peer_effect_pi
 #' }
 #'
 #' @section Values:
@@ -87,6 +94,9 @@
 #'   \item{\code{delta*restricted}}{delta calculated using Oster’s restricted estimator}
 #'   \item{\code{delta_exact}}{delta calculated using correlation-based approach}
 #'   \item{\code{delta_pctbias}}{percent bias when comparing \code{delta*} to \code{delta_exact}}
+#'   \item{\code{delta_sig}}{delta threshold at which focal predictor loses statistical significance at the chosen \code{alpha} (default: 0.05)}
+#'   \item{\code{rxcvGz_sig}}{boundary partial correlation \eqn{r_{X,\mathrm{CV} | Z}} associated with \code{delta_sig}}
+#'   \item{\code{rycvGz_sig}}{boundary partial correlation \eqn{r_{Y,\mathrm{CV} | Z}} associated with \code{delta_sig}}
 #'   \item{\code{var(Y)}}{variance of the dependent variable (\eqn{\sigma_Y^2})}
 #'   \item{\code{var(X)}}{variance of the independent variable (\eqn{\sigma_X^2})}
 #'   \item{\code{var(CV)}}{variance of the confounding variable (\eqn{\sigma_{CV}^2})}
@@ -140,6 +150,14 @@
 #' }
 #' }
 #'
+#' \subsection{RIR for VAM model (beta)}{
+#' \describe{
+#'   \item{\code{RIR}}{Robustness of Inference to Replacement (RIR): number of students needed to be replaced.}
+#'   \item{\code{RIR_perc}}{RIR as \% of students needed to be replaced.}
+#'   \item{\code{peer_effect}}{Peer effect of each replaced student (compared to their replacements) on each of the non-replaced students.}
+#' }
+#' }
+#' 
 #' @note 
 #' For a thoughtful background on benchmark options for ITCV, see 
 #' \href{https://doi.org/10.1111/rssb.12348}{Cinelli & Hazlett (2020)}, 
@@ -179,7 +197,9 @@
 #' # Calculating rxcv and rycv when preserving standard error
 #' pkonfound(est_eff = .5, std_err = .056, n_obs = 6174, eff_thr = .1,
 #'          sdx = 0.22, sdy = 1, R2 = .3, index = "PSE", to_return = "raw_output")
-#' 
+#' # VAM beta
+#' pkonfound(est_eff = 0.14, replace_stu = 0.16, n_obs = 20, eff_thr = 0.15,
+#'           peer_effect_pi = 0.3, index = "VAM")
 #' @export
 #' 
 #' @param est_eff the estimated effect (e.g., an unstandardized beta coefficient or a group mean difference).
@@ -242,7 +262,9 @@
                       to_return = "print",
                       upper_bound = NULL,
                       lower_bound = NULL,
-                      raw_treatment_success = NULL
+                      raw_treatment_success = NULL, 
+                      replace_stu = NULL,
+                      peer_effect_pi = 0.5
                       ) {
   if ("table" %in% to_return) stop("a table can only be
                                    output when using konfound")
@@ -300,7 +322,16 @@
     tails = tails,
     to_return = to_return
   )
-}else if (model_type == "logistic" & !is.null(n_treat)) {
+   } else if (index == "VAM") {
+   out <- test_VAM(
+       est_eff = est_eff,
+       replace_stu = replace_stu,
+       n_obs = n_obs,
+       eff_thr = eff_thr,
+       peer_effect_pi = peer_effect_pi,
+       to_return = to_return
+   )
+   } else if (model_type == "logistic" & !is.null(n_treat)) {
     out <- test_sensitivity_ln(
       est_eff = est_eff,
       std_err = std_err,
@@ -374,9 +405,15 @@ if (!is.null(out)) { # dealing with a strange print issue
 }
 
 if (to_return == "print") {
-  cat("\n")
-  message("For other forms of output, run
-          ?pkonfound and inspect the to_return argument")
+    cat("\n")
+    message("For more information, visit https://konfound-it.org")
+    message(paste0("To explore examples and interpretation tips,\n",
+    "see our Practical Guide at https://konfound-it.org/page/guide/"))
+    cat("\n")
+    message(paste0(
+        "For other forms of output, run\n",
+        "?pkonfound and inspect the to_return argument"
+    ))
 }
 
 message("For models fit in R, consider use of konfound().")