+++ tests/testthat/test-test_pse.R
@@ -78,18 +78,31 @@ test_that("test_pse print output is correct", {
     
     # Expected string
     expected_output <- c(
+        "Component Correlations (fixed standard error):",
         "This function calculates the correlations associated with an omitted confounding ",
         "variable (CV) that generate an estimated effect that is approximately equal to ",
         "the threshold while preserving the originally reported standard error.",
         "",
-        "The correlation between X and CV is 0.214, and the correlation between",
-        "Y and CV is 0.313.",
+        "The correlation between X (focal predictor) and CV is 0.214, and the correlation between",
+        "Y (outcome) and CV is 0.313.",
         "",
         "Conditional on the covariates, the correlation between X and CV is 0.248,",
         "and the correlation between Y and CV is 0.372.",
         "",
         "Including such a CV, the coefficient would change to 0.097, with standard error",
         "of 0.054.",
+        "",
+        "Robustness of Inference to Replacement (fixed standard error)",
+        "To nullify the inference while preserving the reported standard error,",
+        "approximately pi = 0.786 (78.6% of cases) of the data points would need to be replaced.",
+        "This corresponds to replacing about 4854 of 6174 observations.",
+        "",
+        "The replacement cases would need to have a standard deviation of Y equal to 0.96",
+        "in order to maintain the same standard error of beta.",
+        "",
+        "Therefore, if 79% of the cases are replaced with cases for which the effect ",
+        "is modified as above, the estimated effect will be 0.11 with standard error ",
+        "of 0.06 associated with p = 0.05.",
         "",
         "Use to_return = \"raw_output\" to see more specific results."
     )