+++ NEWS.md
@@ -1,3 +1,12 @@
+
+# konfound 1.0.4 (IN DEVELOPMENT)
+
+* add beta version for the VAM model
+* add RIR index calculation for the PSE (component correlation) function
+* include a global message linking to the konfound website and practical guide
+* fix issue where "NULL" appeared at the end of printed output for konfound commands
+* add delta threshold for statistical significance calculation and printed text for COP function
+
 # konfound 1.0.3
 
 * fix sign problems in print output for some cases